+++ .scripts/ci/install.sh
@@ -12,7 +12,7 @@ if [[ "$OS" == "macos-latest" ]]; then
   pip install mpi4py
 elif [[ "$OS" == "ubuntu-latest" ]]; then
   curl -sS https://dl.yarnpkg.com/debian/pubkey.gpg | sudo apt-key add -
-  
+
   sudo apt-key adv --keyserver keyserver.ubuntu.com --recv-keys E298A3A825C0D65DFD57CBB651716619E084DAB9
   sudo add-apt-repository "deb https://cloud.r-project.org/bin/linux/ubuntu $(lsb_release -cs)-cran40/"
   sudo apt-get update -y